Module solana_bucket_map::index_entry

Structs

BucketWithHeader

allocated in contents in a BucketStorage

DataBucketRefCountOccupiedHeader 🔒

header for elements in a bucket needs to be multiple of size_of::()

IndexBucketUsingBitVecBits

allocated in contents in a BucketStorage

IndexEntry

one instance of this per item in the index stored in the index bucket

IndexEntryPlaceInBucket

contains the index of an entry in the index bucket. This type allows us to call methods to interact with the index entry on this type.

MultipleSlots 🔒

required fields when an index element references the data file

PackedRefCount 🔒

hold a big RefCount while leaving room for extra bits to be used for things like ‘Occupied’

PackedStorage 🔒

Pack the storage offset and capacity-when-crated-pow2 fields into a single u64

Enums

OccupiedEnum 🔒

OccupiedEnumTag 🔒

just the values for OccupiedEnum This excludes the contents of any enum value.

Constants

Free__num_enum_0__ 🔒

MAX_LEGAL_REFCOUNT 🔒

63 bits available for ref count

MultipleSlots__num_enum_0__ 🔒

OCCUPIED_FREE 🔒

free, ie. not occupied

OCCUPIED_OCCUPIED 🔒

in use/occupied

OneSlotInIndex__num_enum_0__ 🔒

ZeroSlots__num_enum_0__ 🔒

Type Definitions

DataBucket

IndexBucket

Unions

SingleElementOrMultipleSlots 🔒